+++ CHANGELOG.md
@@ -8,6 +8,22 @@ and this project adheres to
 
 ## [Unreleased]
 
+### [1.22.5] - 2025-08-20
+
+- Improve mobile view of `Manage` page
+
+### [1.22.4] - 2025-08-20
+
+- Fix undername validation regex
+
+### [1.22.3] - 2025-08-20
+
+- Allow controllers to submit transactions
+
+### [1.22.2] - 2025-08-20
+
+- Allow controllers to add undernames
+
 ## [1.22.1] - 2025-08-18
 
 ### Changed